allow a host specifier on the cmd line
 - move k8s specific info inside k8s package
